+++ App/AppStore/CKSoftwareMap+AppLookup.swift
@@ -23,6 +23,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
+import CommerceKit
+
 private var appIdsByName : [String:UInt64]?
 
 extension CKSoftwareMap {