+++ Homebrew/mas.rb
@@ -1,9 +1,11 @@
 class Mas < Formula
   desc "Mac App Store command-line interface"
   homepage "https://github.com/mas-cli/mas"
-  url "https://github.com/mas-cli/mas/archive/v1.4.2.tar.gz"
-  sha256 "f9a751ff84e6dcbaedd4b2ca95b3ca623c739fd3af0b6ca950c321f2ce840bfe"
-  head "https://github.com/mas-cli/mas.git"
+  url "https://github.com/mas-cli/mas.git",
+    :tag => "v1.4.2",
+    :revision => "966872b32820c014a9004691f5da47f170702236",
+    :shallow => true
+  head "https://github.com/mas-cli/mas.git", :shallow => true
 
   bottle do
     cellar :any_skip_relocation
@@ -13,27 +15,14 @@ class Mas < Formula
 
   depends_on :xcode => ["10.0", :build]
 
-  resource "cocoapods" do
-    url "https://dl.bintray.com/phatblat/mas-bottles/master.tar.gz"
-    sha256 "fd8f1b06a2a0276c9005241b45cc19393b7c39cfc91d08da92a307ea2416e966"
-  end
-
   def install
-    # Pre-install a shallow copy of the CocoaPods master repo
-    (buildpath/".brew_home/.cocoapods/repos/master").install resource("cocoapods")
-
-    # Install bundler, then use it to install gems used by project
-    ENV["GEM_HOME"] = buildpath/"gem_home"
-    system "gem", "install", "bundler"
-    ENV.prepend_path "PATH", buildpath/"gem_home/bin"
-    system "bundle", "install"
-    system "bundle", "exec", "pod", "install"
 
     xcodebuild "-workspace", "mas-cli.xcworkspace",
                "-scheme", "mas-cli Release",
                "SYMROOT=#{buildpath.realpath}"
 
     bin.install buildpath/"build/mas"
+    # TODO: Move MasKit.frameworks to prefix/Frameworks
 
     bash_completion.install "contrib/completion/mas-completion.bash" => "mas"
   end