+++ PrivateFrameworks/CommerceKit/CKAccountStore.h
@@ -4,10 +4,12 @@
 //     class-dump is Copyright (C) 1997-1998, 2000-2001, 2004-2013 by Steve Nygard.
 //
 
-// #import "NSObject.h"
+@import Cocoa;
 
-#import "ISStoreURLOperationDelegate-Protocol.h"
-#import <CommerceKit/CKServiceInterface.h>
+//#import "NSObject.h"
+//#import "CKServiceInterface.h"
+
+//#import "ISStoreURLOperationDelegate.h"
 
 @class CKDemoAccount, CKStoreAccount, CKStoreClient, ISStoreAccount, NSArray;
 
@@ -45,6 +47,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property(readonly) NSArray *accounts;
 - (id)accountWithAppleID:(id)arg1;
 - (id)accountForDSID:(id)arg1;
+@property(readonly) BOOL primaryAccountIsPresentAndSignedIn;
 @property(readonly) ISStoreAccount *primaryAccount;
 - (void)removePrimaryAccountObserver:(id)arg1;
 
@@ -59,7 +62,6 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)signOut;
 - (id)storeAccountForAppleID:(id)arg1;
 - (id)storeAccountForDSID:(id)arg1;
-@property(readonly) BOOL primaryAccountIsPresentAndSignedIn;
 @property(readonly) CKStoreAccount *primaryStoreAccount;
 @property(readonly) CKDemoAccount *demoAccount;
 @property(readonly) BOOL isDemoModeEnabled;