+++ PrivateFrameworks/CommerceKit/CKPurchaseController.h
@@ -4,9 +4,9 @@
 //     class-dump is Copyright (C) 1997-1998, 2000-2001, 2004-2013 by Steve Nygard.
 //
 
-#import <CommerceKit/CKServiceInterface.h>
-
-#import <StoreFoundation/SSPurchase.h>
+#import "CKServiceInterface.h"
+//#import <StoreFoundation/SSPurchase.h>
+@import StoreFoundation;
 
 @class NSArray, NSMutableArray, NSNumber;