+++ PrivateFrameworks/CommerceKit/CKServiceInterface.h
@@ -4,6 +4,8 @@
 //     class-dump is Copyright (C) 1997-1998, 2000-2001, 2004-2013 by Steve Nygard.
 //
 
+@import StoreFoundation;
+//#import <StoreFoundation/ISServiceProxy.h>
 #import "ISServiceProxy.h"
 
 @interface CKServiceInterface : ISServiceProxy