+++ PrivateFrameworks/CommerceKit/ISStoreURLOperationDelegate.h
@@ -4,9 +4,9 @@
 //     class-dump is Copyright (C) 1997-1998, 2000-2001, 2004-2013 by Steve Nygard.
 //
 
-//#import "ISURLOperationDelegate.h"
-
-#import "ISURLOperationDelegate-Protocol.h"
+@import StoreFoundation;
+//#import <StoreFoundation/ISURLOperationDelegate.h>
+#import "ISURLOperationDelegate.h"
 
 @class ISStoreURLOperation, NSNumber, NSString;