+++ PrivateFrameworks/StoreFoundation/CKSoftwareProduct.h
@@ -4,6 +4,7 @@
 //     class-dump is Copyright (C) 1997-1998, 2000-2001, 2004-2013 by Steve Nygard.
 //
 
+@import Foundation;
 // #import <Foundation/NSZone.h>
 
 //#import "NSObject.h"