+++ PrivateFrameworks/StoreFoundation/ISURLOperationDelegate.h
@@ -4,9 +4,7 @@
 //     class-dump is Copyright (C) 1997-1998, 2000-2001, 2004-2013 by Steve Nygard.
 //
 
-// #import "ISOperationDelegate.h"
-
-#import "ISOperationDelegate-Protocol.h"
+ #import "ISOperationDelegate.h"
 
 @class ISURLOperation, NSMutableURLRequest, NSURLResponse;